+++ src/Controller/EmailTemplateController.php
@@ -12,6 +12,7 @@ use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpKernel\Attribute\MapQueryString;
 use Symfony\Component\HttpKernel\Attribute\MapRequestPayload;
+use Symfony\Component\HttpKernel\Exception\AccessDeniedHttpException;
 use Symfony\Component\Routing\Attribute\Route;
 use Symfony\Component\Serializer\Encoder\DecoderInterface;
 
@@ -42,6 +43,10 @@ class EmailTemplateController extends AbstractController
     #[Route('/api/templates', methods: ['POST'])]
     public function createTemplate(#[MapRequestPayload] EmailTemplateDto $emailTemplateDto): JsonResponse
     {
+        if (!$this->isGranted('ROLE_EDITOR')) {
+            throw new AccessDeniedHttpException('You don\'t have permission to perform this action!');
+        }
+
         $template = $this->emailTemplateRepository->createEmailTemplate($emailTemplateDto);
 
         return $this->json([
@@ -60,6 +65,10 @@ class EmailTemplateController extends AbstractController
         Request $request,
     ): JsonResponse
     {
+        if (!$this->isGranted('ROLE_EDITOR')) {
+            throw new AccessDeniedHttpException('You don\'t have permission to perform this action!');
+        }
+
         $valuesToChange = $decoder->decode($request->getContent(), 'json');
 
         $template = $this->emailTemplateRepository->updateEmailTemplate($template, $valuesToChange);
@@ -78,6 +87,10 @@ class EmailTemplateController extends AbstractController
         EmailTemplate $template,
     ): JsonResponse
     {
+        if (!$this->isGranted('ROLE_EDITOR')) {
+            throw new AccessDeniedHttpException('You don\'t have permission to perform this action!');
+        }
+
         $this->emailTemplateRepository->deleteTemplate($template);
 
         return $this->json([], 204);
